# api/openapi-lock.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "org.quicko.qpon",
    "description": "Qpon is a coupon framework built by Quicko. Built in a scalable, multi-tenant thought process, the Qpon framework enables businesses to do campaigns and run offers using the Qpon API.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://{{quicko_host}}"
    }
  ],
...
  "paths": {
    "/organizations": {
      "post": {
        "tags": [
          "Organization"
        ],
        "summary": "Create organization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "organization"
                  },
                  "name": {
                    "type": "string"
                  },
                  "currency": {
                    "type": "string"
                  },
                  "theme_colour": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  },
                  "external_id": {
                    "type": "string"
                  }
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "examples": {
                  "200-Example-1": {
                    "value": {
                      "@entity": "org.quicko.qpon.organization",
                      "name": "pro-quicko-qpon-test",
                      "external_id": "org-quicko-qpon-api@quicko.pro"
                    },
                    "summary": "Example 1 for status 200"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Organization"
        ],
        "summary": "Fetch organizations",
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "external_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "filter to get organizations by external id",
            "example": "{{external_id}}"
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page size. Default is 10",
            "example": "10"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Describe the offset",
            "example": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organization_id}": {
      "get": {
        "tags": [
          "Organization"
        ],
        "summary": "Fetch organization",
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Organization id",
            "example": "{{organization_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Organization"
        ],
        "summary": "Update organization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "@entity": "org.quicko.qpon.organization",
                  "name": "com-quicko",
                  "external_id": "org-quicko-qpon-api@quicko.com"
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Organization id",
            "example": "{{organization_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Organization"
        ],
        "summary": "Delete organization",
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Organization id",
            "example": "{{organization_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organization_id}/reports": {
      "get": {
        "tags": [
          "Organization"
        ],
        "summary": "Fetch organization summary",
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/coupon-codes/redeem": {
      "post": {
        "tags": [
          "Redemptions"
        ],
        "summary": "Redeem coupon code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "@entity": "org.quicko.qpon.redemption",
                  "code": "FIRSTITR",
                  "redemption_amount": 250
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "examples": {
                  "200-Example-1": {
                    "value": {
                      "@entity": "org.quicko.qpon.redemption",
                      "code": "FIRSTITR",
                      "redemption_amount": 250
                    },
                    "summary": "Example 1 for status 200"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/redemptions": {
      "get": {
        "tags": [
          "Redemptions"
        ],
        "summary": "Fetch redemptions",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API key for identification",
            "example": "{{api_key}}"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "JWT access token",
            "example": "{{access_token}}"
          },
          {
            "name": "x-api-version",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API version",
            "example": "{{version}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "coupon_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Option filter to get redemptions for a coupon",
            "example": "{{coupon_id}}"
          },
          {
            "name": "campaign_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Option filter to get redemptions for a campaign",
            "example": "{{campaign_id}}"
          },
          {
            "name": "coupon_code_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Option filter to get redemptions for a coupon code",
            "example": "{{coupon_code_id}}"
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Option filter to get redemptions redeemed from time in EPOCH",
            "example": "1716364274000"
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Option filter to get redemptions redeemed to time in EPOCH",
            "example": "1716531482000"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page offset. Default is 0",
            "example": "0"
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page size. Default is 10",
            "example": "10"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
...
  }
}